move `KeycardInteractingEventArgs` to right namespace
